Updated logout component to work with our useUsersStore.
Added logout button to navbar when user is logged in. Also added register button in navbar so that user can reach this page more easily.

Added functionality to show a message saying "You are already logged in..." if user are trying to login or register when they are logged in. (However for this to work properly, ie also when a user types the URL to login/register page in the adress bar, I think we need to store the active user in localStorage).